Split into LegacyProtocol and DatastreamProtocol
