+++ src/Io/DatastreamProtocol.php
@@ -1,0 +1,96 @@
+<?php
+
+namespace Clue\React\Quassel\Io;
+
+use Clue\QDataStream\Writer;
+use Clue\QDataStream\Reader;
+use Clue\QDataStream\QVariant;
+use Clue\QDataStream\Types;
+
+class DatastreamProtocol extends Protocol
+{
+    public function isLegacy()
+    {
+        return false;
+    }
+
+    public function writeVariantList(array $list)
+    {
+        $writer = new Writer(null, $this->types, $this->userTypeWriter);
+
+        // datastream protocol just uses list contents
+        $writer->writeQVariantList($list);
+
+        return (string)$writer;
+    }
+
+    public function writeVariantMap(array $map)
+    {
+        $writer = new Writer(null, $this->types);
+
+        // datastream protocol just uses list contents with UTF-8 keys
+        // the list always starts with a key string, which can be used to tell apart from actual list contents
+        // https://github.com/quassel/quassel/blob/master/src/common/protocols/datastream/datastreampeer.cpp#L80
+        $writer->writeQVariantList($this->mapToList($map));
+
+        return (string)$writer;
+    }
+
+    public function readVariant($packet)
+    {
+        $reader = Reader::fromString($packet, $this->types, $this->userTypeReader);
+
+        // datastrema protocol always uses list contents (even for maps)
+        $value = $reader->readQVariantList();
+
+        // if the first element is a string, then this is actually a map transported as a list
+        // actual lists will always start with an integer request type
+        if (is_string($value[0])) {
+            // datastream protocol uses lists with UTF-8 keys
+            // https://github.com/quassel/quassel/blob/master/src/common/protocols/datastream/datastreampeer.cpp#L109
+            return $this->listToMap($value);
+        }
+
+        if ($value[0] === self::REQUEST_INITDATA) {
+            // first 3 elements are unchanged, everything else should be a map
+            return array_slice($value, 0, 3) + $this->listToMap(array_slice($value, 3));
+        }
+
+        return $value;
+    }
+
+    /**
+     * converts the given map to a list
+     *
+     * @param mixed[]|array<mixed> $map
+     * @return mixed[]|array<mixed>
+     * @internal
+     */
+    public function mapToList($map)
+    {
+        $list = array();
+        foreach ($map as $key => $value) {
+            // explicitly pass key as UTF-8 byte array
+            // pass value with automatic type detection
+            $list []= new QVariant($key, Types::TYPE_QBYTE_ARRAY);
+            $list []= $value;
+        }
+        return $list;
+    }
+
+    /**
+     * converts the given list to a map
+     *
+     * @param mixed[]|array<mixed> $list
+     * @return mixed[]|array<mixed>
+     * @internal
+     */
+    public function listToMap($list)
+    {
+        $map = array();
+        for ($i = 0, $n = count($list); $i < $n; $i += 2) {
+            $map[$list[$i]] = $list[$i + 1];
+        }
+        return $map;
+    }
+}
